requirements.txt
bson==0.5.10
requests==2.32.3
pre-commit==3.8.0
pytz==2024.1
python-dotenv==1.0.1
pymongo==4.8.0
